forecasting/pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["senkovskiy"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
pandas = "2.2.3"
plotly = "5.24.1"
python-dotenv = "^1.0.1"
ipywidgets = "^8.1.5"
holidays = "0.57"
mlflow = "^2.17.2"
deltalake = "^0.21.0"
duckdb = "^1.1.3"
sktime = "^0.34.0"
optuna = "^4.0.0"
catboost = "^1.2.7"
numpy = "1.26.4"
seaborn = "^0.13.2"
clickhouse-connect = "^0.8.6"
flake8 = "^7.1.1"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"
notebook = "^7.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
